so buttons fertig?

+++ src/scripts/prepareRolecallChannel.ts
@@ -1,5 +1,6 @@
 import {MessageActionRow, MessageButton, MessageEmbed, MessageSelectMenu, SnowflakeUtil} from "discord.js";
 import Helper from "../helpers/helper";
+import {PRINT_ROLECALL} from "../config/envConfig";
 
 export async function prepareRolecallChannel(helper: Helper) : Promise<void> {
 
@@ -16,51 +17,53 @@ export async function prepareRolecallChannel(helper: Helper) : Promise<void> {
         });
         if (!messages) return;
         await messages.forEach(message => {
+            if (message.embeds.length > 0) return;
             if (!message.deleted && message.deletable && message.id) message.delete();
         });
     } catch (e) {}
 
     // PRINT the interaction buttons
-    const row1 = new MessageActionRow();
-    const row2 = new MessageActionRow();
-    const row3 = new MessageActionRow();
+    if(PRINT_ROLECALL) {
+        const row1 = new MessageActionRow();
+        const row2 = new MessageActionRow();
+        const row3 = new MessageActionRow();
 
-    let bsc_buttons = [
-        new MessageButton().setCustomId('semester_sem1').setLabel('BSc. 1').setStyle('PRIMARY'),
-        new MessageButton().setCustomId('semester_sem2').setLabel('BSc. 2').setStyle('PRIMARY'),
-        new MessageButton().setCustomId('semester_sem3').setLabel('BSc. 3').setStyle('PRIMARY'),
-        new MessageButton().setCustomId('semester_sem4').setLabel('BSc. 4').setStyle('PRIMARY'),
-        new MessageButton().setCustomId('semester_sem5').setLabel('BSc. 5+').setStyle('PRIMARY'),
-    ];
+        let bsc_buttons = [
+            new MessageButton().setCustomId('semester_sem1').setLabel('BSc. 1').setStyle('PRIMARY'),
+            new MessageButton().setCustomId('semester_sem2').setLabel('BSc. 2').setStyle('PRIMARY'),
+            new MessageButton().setCustomId('semester_sem3').setLabel('BSc. 3').setStyle('PRIMARY'),
+            new MessageButton().setCustomId('semester_sem4').setLabel('BSc. 4').setStyle('PRIMARY'),
+            new MessageButton().setCustomId('semester_sem5').setLabel('BSc. 5+').setStyle('PRIMARY'),
+        ];
 
-    bsc_buttons.forEach(btn => row1.addComponents(btn));
-    row2.addComponents(new MessageButton().setCustomId('master_master').setLabel('Master').setStyle('PRIMARY').setDisabled(true))
-    const embed1 = new MessageEmbed()
-        .setColor('#0099ff')
-        .setTitle('BSc, MSc, OmG')
-        .setDescription('Joa, klick mal dahin wo du dich grade so siehst:');
-    channel?.send({ embeds: [embed1], components: [row1, row2] })
+        bsc_buttons.forEach(btn => row1.addComponents(btn));
+        row2.addComponents(new MessageButton().setCustomId('master_master').setLabel('Master').setStyle('PRIMARY'));
+        const embed1 = new MessageEmbed()
+            .setColor('#0099ff')
+            .setTitle('BSc, MSc, OmG')
+            .setDescription('Joa, klick mal dahin wo du dich grade so siehst:');
+        channel?.send({ embeds: [embed1], components: [row1, row2] })
 
 
 
-    let vertiefungsRichtungen: { label: string, value: string }[] = [
-        {value: 'Vertiefung Bauingenieurwesen', label: 'Bauingenieurwesen'},
-        {value: 'Vertiefung Computational_Robotics', label: 'Computational_Robotics'},
-        {value: 'Vertiefung Elektrotechnik-Informationstechnik', label: 'Elektrotechnik-Informationstechnik'},
-        {value: 'Vertiefung Informatik', label: 'Informatik'},
-        {value: 'Vertiefung Maschinenbau', label: 'Maschinenbau'},
-        {value: 'Vertiefung Mathe-Mechanik', label: 'Mathe-Mechanik'},
-        {value: 'Vertiefung Strömung-Verbrennung', label: 'Strömung-Verbrennung'},
-    ];
-    row3.addComponents(new MessageSelectMenu()
-        .setCustomId('vertiefung_SELECT')
-        .setPlaceholder('Nothing selected')
-        .addOptions(vertiefungsRichtungen)
-        .setMinValues(1).setMaxValues(1));
-    const embed2 = new MessageEmbed()
-        .setColor('#0099ff')
-        .setTitle('Die Qual der Wahl')
-        .setDescription('Im 4 BSc Semester, sowie im Master wählst du eine Vertiefungsrichtung. Du kannst unten diese auswählen:');
-    channel?.send({ embeds: [embed2], components: [row3] })
-
+        let vertiefungsRichtungen: { label: string, value: string }[] = [
+            {value: 'Vertiefung Bauingenieurwesen', label: 'Bauingenieurwesen'},
+            {value: 'Vertiefung Computational_Robotics', label: 'Computational_Robotics'},
+            {value: 'Vertiefung Elektrotechnik-Informationstechnik', label: 'Elektrotechnik-Informationstechnik'},
+            {value: 'Vertiefung Informatik', label: 'Informatik'},
+            {value: 'Vertiefung Maschinenbau', label: 'Maschinenbau'},
+            {value: 'Vertiefung Mathe-Mechanik', label: 'Mathe-Mechanik'},
+            {value: 'Vertiefung Strömung-Verbrennung', label: 'Strömung-Verbrennung'},
+        ];
+        row3.addComponents(new MessageSelectMenu()
+            .setCustomId('vertiefung_SELECT')
+            .setPlaceholder('Nothing selected')
+            .addOptions(vertiefungsRichtungen)
+            .setMinValues(1).setMaxValues(1));
+        const embed2 = new MessageEmbed()
+            .setColor('#0099ff')
+            .setTitle('Die Qual der Wahl')
+            .setDescription('Im 4 BSc Semester, sowie im Master wählst du eine Vertiefungsrichtung. Du kannst unten diese auswählen:');
+        channel?.send({ embeds: [embed2], components: [row3] })
+    }
 }
